+++ src/main/java/insane96mcp/enhancedai/mixin/WalkNodeEvaluatorMixin.java
@@ -5,6 +5,7 @@ import com.llamalad7.mixinextras.sugar.Local;
 import insane96mcp.enhancedai.module.mobs.Pathfinding;
 import net.minecraft.world.level.block.TrapDoorBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.pathfinder.PathType;
 import net.minecraft.world.level.pathfinder.WalkNodeEvaluator;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -17,5 +18,12 @@ public class WalkNodeEvaluatorMixin {
                 || !original)
             return original;
         return !blockstate.getValue(TrapDoorBlock.OPEN);
+    }
+
+    @ModifyExpressionValue(method = "getPathTypeWithinMobBB", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/pathfinder/WalkNodeEvaluator;getPathType(Lnet/minecraft/world/level/pathfinder/PathfindingContext;III)Lnet/minecraft/world/level/pathfinder/PathType;", ordinal = 2))
+    private PathType enhancedai$allowWalkingOnRails(PathType original) {
+        if (!Pathfinding.shouldMobsWalkOnRails())
+            return original;
+        return PathType.RAIL;
     }
 }